Add files via upload
IMDb Data Analysis: Exploring Movie Datasets and Insights.

Short Description:

In this repository, you'll find a comprehensive analysis of IMDb movie data to uncover interesting trends, insights, and patterns in the film industry. Whether you're a movie enthusiast or a data analyst this project decodes the inner matter.

Disclaimer;
As a fresher I have implemented whatever I learnt, so any corrections or mistake are found in my analysis, they are taken in a good attitude.